+++ app/Http/Controllers/MemberController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\UpdateMemberRequest;
 use App\Http\Resources\MemberResource;
 use App\Http\Resources\MemberRolesResource;
 use App\Models\Organization;
 use App\Models\OrganizationInvitation;
 use App\Models\OrganizationMember;
+use App\Models\Role;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
@@ -14,27 +16,23 @@ use Inertia\Inertia;
 class MemberController extends Controller
 {
     /**
-     * Create the controller instance.
-     */
-    public function __construct()
-    {
-        $this->authorizeresource(organizationmember::class, 'member');
-    }
-
-    /**
      * Display a listing of the resource.
      */
     public function index(Request $request)
     {
-        $organization = Organization::find(session('organization_id'));
+        $this->authorize('viewAny', OrganizationMember::class);
 
-        $roles = $organization->roles()->orderBy('name')->select(['id', 'name'])->get();
+        $organizationId = session('organization_id');
+        $organization = Organization::find($organizationId);
 
-        // If the user is searching for a member, return the search results.
-        if ($request->input('query')) {
+        $roles = Role::where('organization_id', $organizationId)
+            ->orderBy('name')
+            ->select(['id', 'name'])
+            ->get();
+
+        if ($request->filled('query')) {
             $searchResults = User::search($request->input('query'))
-                ->orderBy('first_name')
-                ->whereIn('id', $organization->members->pluck('id')->toArray())
+                ->whereIn('id', $organization->members()->pluck('users.id')->toArray())
                 ->paginate(10);
 
             return Inertia::render('Members', [
@@ -43,12 +41,10 @@ class MemberController extends Controller
             ]);
         }
 
-        // Otherwise, return the members of the organization with their memberships.
-        $membersPagination = $organization
-                ->members()
-                ->orderBy('first_name')
-                ->orderBy('users.id')
-                ->cursorPaginate(10, ['users.id', 'users.first_name', 'users.last_name', 'users.email']);
+        $membersPagination = $organization->members()
+            ->orderBy('first_name')
+            ->orderBy('users.id')
+            ->paginate(10);
 
         return Inertia::render('Members', [
             'pagination' => MemberResource::collection($membersPagination),
@@ -56,19 +52,20 @@ class MemberController extends Controller
         ]);
     }
 
-    public function update(Request $request, OrganizationMember $member)
+    /**
+     * Update the specified resource in storage.
+     */
+    public function update(UpdateMemberRequest $request, OrganizationMember $member)
     {
-        $member->update($request->validate([
-            'role_id' => 'required|exists:roles,id',
-        ]));
+        $this->authorize('update', $member);
 
-        $organization = Organization::find(session('organization_id'));
+        $member->update($request->validated());
 
-        $role = $organization->roles()->find($request->input('role_id'));
+        $role = Role::find($request->input('role_id'));
 
         $member->user->syncRoles($role->name);
 
-        return back()->with(['message' => 'Member updated successfully.', 'type' => 'success']);
+        return back()->with(['message' => 'Member updated successfully!', 'type' => 'success']);
     }
 
     /**
@@ -76,25 +73,25 @@ class MemberController extends Controller
      */
     public function destroy(OrganizationMember $member)
     {
+        $this->authorize('delete', $member);
+
         // If the user is the owner of the organization, return an error.
         if ($member->organization->user_id === $member->user_id) {
             return back()->with(['message' => 'You cannot remove the owner of the organization.', 'type' => 'failure']);
         }
 
         // Get the invitation for the member.
-        $invitation = OrganizationInvitation::where('user_id', $member->user_id)->where('organization_id', $member->organization_id)->first();
-
-        if ($invitation) {
-            $invitation->delete();
-        }
+        OrganizationInvitation::where('user_id', $member->user_id)
+            ->where('organization_id', $member->organization_id)
+            ->delete();
 
         $member->delete();
 
         // If the member is the current user, forget the organization_id session variable.
-        if ($member->user_id === auth()->user()->id) {
+        if ($member->user_id === auth()->id()) {
             session()->forget('organization_id');
         }
 
-        return back()->with(['message' => 'Member deleted successfully.', 'type' => 'success']);
+        return back()->with(['message' => 'Member deleted successfully!', 'type' => 'success']);
     }
 }
